+++ packages/server/graphql/mutations/removePokerTemplateScaleValue.ts
@@ -1,7 +1,5 @@
 import {GraphQLID, GraphQLNonNull, GraphQLString} from 'graphql'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
-import getRethink from '../../database/rethinkDriver'
-import {RValue} from '../../database/stricterR'
 import getKysely from '../../postgres/getKysely'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import publish from '../../utils/publish'
@@ -25,15 +23,13 @@ const removePokerTemplateScaleValue = {
     {scaleId, label}: {scaleId: string; label: string},
     {authToken, dataLoader, socketId: mutatorId}: GQLContext
   ) {
-    const r = await getRethink()
     const pg = getKysely()
-    const now = new Date()
     const operationId = dataLoader.share()
     const subOptions = {operationId, mutatorId}
     const viewerId = getUserId(authToken)
 
     // AUTH
-    const scale = await r.table('TemplateScale').get(scaleId).run()
+    const scale = await dataLoader.get('templateScales').load(scaleId)
     if (!scale || scale.removedAt) {
       return standardError(new Error('Did not find an active scale'), {userId: viewerId})
     }
@@ -51,27 +47,12 @@ const removePokerTemplateScaleValue = {
     }
 
     // RESOLUTION
-    await r
-      .table('TemplateScale')
-      .get(scaleId)
-      .update((row: RValue) => ({
-        values: row('values').deleteAt(row('values').offsetsOf(oldScaleValue).nth(0)),
-        updatedAt: now
-      }))
-      .run()
-
-    // mark all templates using this scale as updated
-    const updatedDimensions = await r
-      .table('TemplateDimension')
-      .getAll(scaleId, {index: 'scaleId'})
-      .run()
-    const updatedTemplateIds = updatedDimensions.map(({templateId}) => templateId)
     await pg
-      .updateTable('MeetingTemplate')
-      .set({updatedAt: now})
-      .where('id', 'in', updatedTemplateIds)
+      .deleteFrom('TemplateScaleValue')
+      .where('templateScaleId', '=', scaleId)
+      .where('label', '=', label)
       .execute()
-
+    dataLoader.clearAll('templateScales')
     const data = {scaleId}
     publish(
       SubscriptionChannel.TEAM,