+++ packages/server/graphql/public/types/RemoveTeamMemberIntegrationAuthSuccess.ts
@@ -10,7 +10,7 @@ export type RemoveTeamMemberIntegrationAuthSuccessSource = {
 const RemoveTeamMemberIntegrationAuthSuccess: RemoveTeamMemberIntegrationAuthSuccessResolvers = {
   teamMember: ({teamId, userId}, _args, {dataLoader}) => {
     const teamMemberId = toTeamMemberId(teamId, userId)
-    return dataLoader.get('teamMembers').load(teamMemberId)
+    return dataLoader.get('teamMembers').loadNonNull(teamMemberId)
   },
   user: async ({userId}, _args, {dataLoader}) => {
     return dataLoader.get('users').loadNonNull(userId)